+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
--   UserBlock a11y: remove avatar keyboard focus to avoid the redundant keyboard navigation with the name element
+-   UserBlock a11y: remove avatar keyboard focus to avoid the redundant keyboard navigation with the name element.
+-   Dialog & Lightbox: fixed focus trap on `disabled=false` elements.
 
 ## [2.2.18][] - 2022-05-20